adxl powerUp, powerDown implemented. CommController changed to singleton. XML is parsed as bytes not integers now

+++ source/src/paddlefish/comm/I2CComm.java
@@ -8,22 +8,22 @@ import java.util.Objects;
 public class I2CComm {
 	// Count of available device addresses
 	private int devAddrCnt;
-	// Integer: device address (Ex: read as hexadecimal, but kept as decimal value)
+	// Byte: Device address 
 	// Boolean: If the address is active
-	private HashMap<Integer,Boolean> devAddrInf;
+	private HashMap<Byte,Boolean> devAddrInf;
 	
 	// Default constructor
 	public I2CComm()
 	{
 		devAddrCnt = 0;
-		devAddrInf = new HashMap<Integer,Boolean>();
+		devAddrInf = new HashMap<Byte,Boolean>();
 	}
 
 	// Initialize devAddCnt already
 	public I2CComm(int devAddrCnt)
 	{
 		this.devAddrCnt = devAddrCnt;
-		devAddrInf = new HashMap<Integer,Boolean>();
+		devAddrInf = new HashMap<Byte,Boolean>();
 	}
 	
 	public int getDevAddrCnt() {
@@ -34,15 +34,15 @@ public class I2CComm {
 		this.devAddrCnt = devAddrCnt;
 	}
 
-	public HashMap<Integer, Boolean> getDevAddrInf() {
+	public HashMap<Byte, Boolean> getDevAddrInf() {
 		return devAddrInf;
 	}
 
-	public void setDevAddrInf(HashMap<Integer, Boolean> devAddrInf) {
+	public void setDevAddrInf(HashMap<Byte, Boolean> devAddrInf) {
 		this.devAddrInf = devAddrInf;
 	}
 	
-	public void addDeviceAddr(int addr, boolean isActive)
+	public void addDeviceAddr(Byte addr, boolean isActive)
 	{
 		// If map is already full do not add
 		if((this.devAddrCnt==this.devAddrInf.size())&&
@@ -57,11 +57,11 @@ public class I2CComm {
 	}
 	
 	/* Only one of the device addresses is active. Get the adress of the active one */
-	public Integer getActiveDeviceAddr()
+	public Byte getActiveDeviceAddr()
 	{
 		if(this.devAddrInf!=null)
 		{
-		    for (Entry<Integer, Boolean> entry : devAddrInf.entrySet()) 
+		    for (Entry<Byte, Boolean> entry : devAddrInf.entrySet()) 
 		    {
 		        if (Objects.equals(true, entry.getValue())) 
 		        {
